backend/requirements.txt
ariadne==0.11.0
asgiref==3.2.7
astroid==2.4.0
Babel==2.8.0
click==7.1.2
Django==3.0.5
django-phonenumber-field==4.0.0
graphql-core==3.0.4
h11==0.9.0
httptools==0.1.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumberslite==8.12.2
pkg-resources==0.0.0
pylint==2.5.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
starlette==0.13.4
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
wrapt==1.12.1
